+++ extensions/mongodb/Connection.php
@@ -148,7 +148,7 @@ class Connection extends Component
         if ($this->defaultDatabaseName === null) {
             if (isset($this->options['db'])) {
                 $this->defaultDatabaseName = $this->options['db'];
-            } elseif (preg_match('/^mongodb:\\/\\/.+\\/(.+)$/s', $this->dsn, $matches)) {
+            } elseif (preg_match('/^mongodb:\\/\\/.+\\/([^?&]+)/s', $this->dsn, $matches)) {
                 $this->defaultDatabaseName = $matches[1];
             } else {
                 throw new InvalidConfigException("Unable to determine default database name from dsn.");