+++ extensions/sphinx/ActiveQuery.php
@@ -73,7 +73,7 @@ use yii\db\ActiveRelationTrait;
  * A relation is specified by [[link]] which represents the association between columns
  * of different tables; and the multiplicity of the relation is indicated by [[multiple]].
  *
- * If a relation involves a pivot table, it may be specified by [[via()]].
+ * If a relation involves a junction table, it may be specified by [[via()]].
  * This methods may only be called in a relational context. Same is true for [[inverseOf()]], which
  * marks a relation as inverse of another relation.
  *
@@ -225,7 +225,7 @@ class ActiveQuery extends Query implements ActiveQueryInterface
             // lazy loading a relational query
             if ($this->via instanceof self) {
                 // via pivot index
-                $viaModels = $this->via->findPivotRows([$this->primaryModel]);
+                $viaModels = $this->via->findJunctionRows([$this->primaryModel]);
                 $this->filterByModels($viaModels);
             } elseif (is_array($this->via)) {
                 // via relation