+++ extensions/sphinx/Query.php
@@ -8,10 +8,9 @@
 namespace yii\sphinx;
 
 use Yii;
-use yii\base\Component;
 use yii\base\InvalidCallException;
+use yii\base\NotSupportedException;
 use yii\db\Expression;
-use yii\db\QueryInterface;
 use yii\db\QueryTrait;
 
 /**
@@ -46,31 +45,8 @@ use yii\db\QueryTrait;
  * @author Paul Klimov <klimov.paul@gmail.com>
  * @since 2.0
  */
-class Query extends Component implements QueryInterface
+class Query extends \yii\db\Query
 {
-    use QueryTrait;
-
-    /**
-     * @var array the columns being selected. For example, `['id', 'group_id']`.
-     * This is used to construct the SELECT clause in a SQL statement. If not set, if means selecting all columns.
-     * @see select()
-     */
-    public $select;
-    /**
-     * @var string additional option that should be appended to the 'SELECT' keyword.
-     */
-    public $selectOption;
-    /**
-     * @var boolean whether to select distinct rows of data only. If this is set true,
-     * the SELECT clause would be changed to SELECT DISTINCT.
-     */
-    public $distinct;
-    /**
-     * @var array the index(es) to be selected from. For example, `['idx_user', 'idx_user_delta']`.
-     * This is used to construct the FROM clause in a SQL statement.
-     * @see from()
-     */
-    public $from;
     /**
      * @var string|Expression text, which should be searched in fulltext mode.
      * This value will be composed into MATCH operator inside the WHERE clause.
@@ -79,11 +55,6 @@ class Query extends Component implements QueryInterface
      * see [[match()]] for details.
      */
     public $match;
-    /**
-     * @var array how to group the query results. For example, `['company', 'department']`.
-     * This is used to construct the GROUP BY clause in a SQL statement.
-     */
-    public $groupBy;
     /**
      * @var string WITHIN GROUP ORDER BY clause. This is a Sphinx specific extension
      * that lets you control how the best row within a group will to be selected.
@@ -96,11 +67,6 @@ class Query extends Component implements QueryInterface
      * that lets you control a number of per-query options.
      */
     public $options;
-    /**
-     * @var array list of query parameter values indexed by parameter placeholders.
-     * For example, `[':name' => 'Dan', ':age' => 31]`.
-     */
-    public $params = [];
     /**
      * @var callable PHP callback, which should be used to fetch source data for the snippets.
      * Such callback will receive array of query result rows as an argument and must return the
@@ -165,222 +131,38 @@ class Query extends Component implements QueryInterface
 
     /**
      * Creates a Sphinx command that can be used to execute this query.
-     * @param Connection $connection the Sphinx connection used to generate the SQL statement.
+     * @param Connection $db the Sphinx connection used to generate the SQL statement.
      * If this parameter is not given, the `sphinx` application component will be used.
      * @return Command the created Sphinx command instance.
      */
-    public function createCommand($connection = null)
+    public function createCommand($db = null)
     {
-        $this->setConnection($connection);
-        $connection = $this->getConnection();
-        list ($sql, $params) = $connection->getQueryBuilder()->build($this);
+        $this->setConnection($db);
+        $db = $this->getConnection();
+        list ($sql, $params) = $db->getQueryBuilder()->build($this);
 
-        return $connection->createCommand($sql, $params);
+        return $db->createCommand($sql, $params);
     }
 
     /**
-     * Executes the query and returns all results as an array.
-     * @param Connection $db the Sphinx connection used to generate the SQL statement.
-     * If this parameter is not given, the `sphinx` application component will be used.
-     * @return array the query results. If the query results in nothing, an empty array will be returned.
+     * @inheritdoc
      */
-    public function all($db = null)
+    public function populate($rows)
     {
-        $rows = $this->createCommand($db)->queryAll();
-        $rows = $this->fillUpSnippets($rows);
-        if ($this->indexBy === null) {
-            return $rows;
-        }
-        $result = [];
-        foreach ($rows as $row) {
-            if (is_string($this->indexBy)) {
-                $key = $row[$this->indexBy];
-            } else {
-                $key = call_user_func($this->indexBy, $row);
-            }
-            $result[$key] = $row;
-        }
-
-        return $result;
+        return parent::populate($this->fillUpSnippets($rows));
     }
 
     /**
-     * Executes the query and returns a single row of result.
-     * @param Connection $db the Sphinx connection used to generate the SQL statement.
-     * If this parameter is not given, the `sphinx` application component will be used.
-     * @return array|boolean the first row (in terms of an array) of the query result. False is returned if the query
-     * results in nothing.
+     * @inheritdoc
      */
     public function one($db = null)
     {
-        $row = $this->createCommand($db)->queryOne();
+        $row = parent::one($db);
         if ($row !== false) {
             list ($row) = $this->fillUpSnippets([$row]);
         }
 
         return $row;
-    }
-
-    /**
-     * Returns the query result as a scalar value.
-     * The value returned will be the first column in the first row of the query results.
-     * @param Connection $db the Sphinx connection used to generate the SQL statement.
-     * If this parameter is not given, the `sphinx` application component will be used.
-     * @return string|boolean the value of the first column in the first row of the query result.
-     * False is returned if the query result is empty.
-     */
-    public function scalar($db = null)
-    {
-        return $this->createCommand($db)->queryScalar();
-    }
-
-    /**
-     * Executes the query and returns the first column of the result.
-     * @param Connection $db the Sphinx connection used to generate the SQL statement.
-     * If this parameter is not given, the `sphinx` application component will be used.
-     * @return array the first column of the query result. An empty array is returned if the query results in nothing.
-     */
-    public function column($db = null)
-    {
-        return $this->createCommand($db)->queryColumn();
-    }
-
-    /**
-     * Returns the number of records.
-     * @param string $q the COUNT expression. Defaults to '*'.
-     * Make sure you properly quote column names in the expression.
-     * @param Connection $db the Sphinx connection used to generate the SQL statement.
-     * If this parameter is not given, the `sphinx` application component will be used.
-     * @return integer number of records
-     */
-    public function count($q = '*', $db = null)
-    {
-        $this->select = ["COUNT($q)"];
-
-        return $this->createCommand($db)->queryScalar();
-    }
-
-    /**
-     * Returns the sum of the specified column values.
-     * @param string $q the column name or expression.
-     * Make sure you properly quote column names in the expression.
-     * @param Connection $db the Sphinx connection used to generate the SQL statement.
-     * If this parameter is not given, the `sphinx` application component will be used.
-     * @return integer the sum of the specified column values
-     */
-    public function sum($q, $db = null)
-    {
-        $this->select = ["SUM($q)"];
-
-        return $this->createCommand($db)->queryScalar();
-    }
-
-    /**
-     * Returns the average of the specified column values.
-     * @param string $q the column name or expression.
-     * Make sure you properly quote column names in the expression.
-     * @param Connection $db the Sphinx connection used to generate the SQL statement.
-     * If this parameter is not given, the `sphinx` application component will be used.
-     * @return integer the average of the specified column values.
-     */
-    public function average($q, $db = null)
-    {
-        $this->select = ["AVG($q)"];
-
-        return $this->createCommand($db)->queryScalar();
-    }
-
-    /**
-     * Returns the minimum of the specified column values.
-     * @param string $q the column name or expression.
-     * Make sure you properly quote column names in the expression.
-     * @param Connection $db the Sphinx connection used to generate the SQL statement.
-     * If this parameter is not given, the `sphinx` application component will be used.
-     * @return integer the minimum of the specified column values.
-     */
-    public function min($q, $db = null)
-    {
-        $this->select = ["MIN($q)"];
-
-        return $this->createCommand($db)->queryScalar();
-    }
-
-    /**
-     * Returns the maximum of the specified column values.
-     * @param string $q the column name or expression.
-     * Make sure you properly quote column names in the expression.
-     * @param Connection $db the Sphinx connection used to generate the SQL statement.
-     * If this parameter is not given, the `sphinx` application component will be used.
-     * @return integer the maximum of the specified column values.
-     */
-    public function max($q, $db = null)
-    {
-        $this->select = ["MAX($q)"];
-
-        return $this->createCommand($db)->queryScalar();
-    }
-
-    /**
-     * Returns a value indicating whether the query result contains any row of data.
-     * @param Connection $db the Sphinx connection used to generate the SQL statement.
-     * If this parameter is not given, the `sphinx` application component will be used.
-     * @return boolean whether the query result contains any row of data.
-     */
-    public function exists($db = null)
-    {
-        $this->select = [new Expression('1')];
-
-        return $this->scalar($db) !== false;
-    }
-
-    /**
-     * Sets the SELECT part of the query.
-     * @param string|array $columns the columns to be selected.
-     * Columns can be specified in either a string (e.g. "id, name") or an array (e.g. ['id', 'name']).
-     * The method will automatically quote the column names unless a column contains some parenthesis
-     * (which means the column contains a Sphinx expression).
-     * @param string $option additional option that should be appended to the 'SELECT' keyword.
-     * @return static the query object itself
-     */
-    public function select($columns, $option = null)
-    {
-        if (!is_array($columns)) {
-            $columns = preg_split('/\s*,\s*/', trim($columns), -1, PREG_SPLIT_NO_EMPTY);
-        }
-        $this->select = $columns;
-        $this->selectOption = $option;
-
-        return $this;
-    }
-
-    /**
-     * Sets the value indicating whether to SELECT DISTINCT or not.
-     * @param boolean $value whether to SELECT DISTINCT or not.
-     * @return static the query object itself
-     */
-    public function distinct($value = true)
-    {
-        $this->distinct = $value;
-
-        return $this;
-    }
-
-    /**
-     * Sets the FROM part of the query.
-     * @param string|array $tables the table(s) to be selected from. This can be either a string (e.g. `'idx_user'`)
-     * or an array (e.g. `['idx_user', 'idx_user_delta']`) specifying one or several index names.
-     * The method will automatically quote the table names unless it contains some parenthesis
-     * (which means the table is given as a sub-query or Sphinx expression).
-     * @return static the query object itself
-     */
-    public function from($tables)
-    {
-        if (!is_array($tables)) {
-            $tables = preg_split('/\s*,\s*/', trim($tables), -1, PREG_SPLIT_NO_EMPTY);
-        }
-        $this->from = $tables;
-
-        return $this;
     }
 
     /**
@@ -405,210 +187,35 @@ class Query extends Component implements QueryInterface
     }
 
     /**
-     * Sets the WHERE part of the query.
-     *
-     * The method requires a $condition parameter, and optionally a $params parameter
-     * specifying the values to be bound to the query.
-     *
-     * The $condition parameter should be either a string (e.g. 'id=1') or an array.
-     * If the latter, it must be in one of the following two formats:
-     *
-     * - hash format: `['column1' => value1, 'column2' => value2, ...]`
-     * - operator format: `[operator, operand1, operand2, ...]`
-     *
-     * A condition in hash format represents the following SQL expression in general:
-     * `column1=value1 AND column2=value2 AND ...`. In case when a value is an array or a Query object,
-     * an `IN` expression will be generated. And if a value is null, `IS NULL` will be used
-     * in the generated expression. Below are some examples:
-     *
-     * - `['type' => 1, 'status' => 2]` generates `(type = 1) AND (status = 2)`.
-     * - `['id' => [1, 2, 3], 'status' => 2]` generates `(id IN (1, 2, 3)) AND (status = 2)`.
-     * - `['status' => null] generates `status IS NULL`.
-     * - `['id' => $query]` generates `id IN (...sub-query...)`
-     *
-     * A condition in operator format generates the SQL expression according to the specified operator, which
-     * can be one of the followings:
-     *
-     * - `and`: the operands should be concatenated together using `AND`. For example,
-     *   `['and', 'id=1', 'id=2']` will generate `id=1 AND id=2`. If an operand is an array,
-     *   it will be converted into a string using the rules described here. For example,
-     *   `['and', 'type=1', ['or', 'id=1', 'id=2']]` will generate `type=1 AND (id=1 OR id=2)`.
-     *   The method will NOT do any quoting or escaping.
-     *
-     * - `or`: similar to the `and` operator except that the operands are concatenated using `OR`.
-     *
-     * - `between`: operand 1 should be the column name, and operand 2 and 3 should be the
-     *   starting and ending values of the range that the column is in.
-     *   For example, `['between', 'id', 1, 10]` will generate `id BETWEEN 1 AND 10`.
-     *
-     * - `not between`: similar to `between` except the `BETWEEN` is replaced with `NOT BETWEEN`
-     *   in the generated condition.
-     *
-     * - `in`: operand 1 should be a column or DB expression with parenthesis. Operand 2 can be an array
-     *   or a Query object. If the former, the array represents the range of the values that the column
-     *   or DB expression should be in. If the latter, a sub-query will be generated to represent the range.
-     *   For example, `['in', 'id', [1, 2, 3]]` will generate `id IN (1, 2, 3)`;
-     *   `['in', 'id', (new Query)->select('id')->from('user'))]` will generate
-     *   `id IN (SELECT id FROM user)`. The method will properly quote the column name and escape values in the range.
-     *   The `in` operator also supports composite columns. In this case, operand 1 should be an array of the columns,
-     *   while operand 2 should be an array of arrays or a `Query` object representing the range of the columns.
-     *
-     * - `not in`: similar to the `in` operator except that `IN` is replaced with `NOT IN` in the generated condition.
-     *
-     * - `like`: operand 1 should be a column or DB expression, and operand 2 be a string or an array representing
-     *   the values that the column or DB expression should be like.
-     *   For example, `['like', 'name', '%tester%']` will generate `name LIKE '%tester%'`.
-     *   When the value range is given as an array, multiple `LIKE` predicates will be generated and concatenated
-     *   using `AND`. For example, `['like', 'name', ['%test%', '%sample%']]` will generate
-     *   `name LIKE '%test%' AND name LIKE '%sample%'`.
-     *   The method will properly quote the column name and escape values in the range.
-     *   Sometimes, you may want to add the percentage characters to the matching value by yourself, you may supply
-     *   a third operand `false` to do so. For example, `['like', 'name', '%tester', false]` will generate `name LIKE '%tester'`.
-     *
-     * - `or like`: similar to the `like` operator except that `OR` is used to concatenate the `LIKE`
-     *   predicates when operand 2 is an array.
-     *
-     * - `not like`: similar to the `like` operator except that `LIKE` is replaced with `NOT LIKE`
-     *   in the generated condition.
-     *
-     * - `or not like`: similar to the `not like` operator except that `OR` is used to concatenate
-     *   the `NOT LIKE` predicates.
-     *
-     * @param string|array $condition the conditions that should be put in the WHERE part.
-     * @param array $params the parameters (name => value) to be bound to the query.
-     * @return static the query object itself
-     * @see andWhere()
-     * @see orWhere()
+     * @inheritdoc
      */
-    public function where($condition, $params = [])
+    public function join($type, $table, $on = '', $params = [])
     {
-        $this->where = $condition;
-        $this->addParams($params);
-        return $this;
+        throw new NotSupportedException('"' . __METHOD__ . '" is not supported.');
     }
 
     /**
-     * Adds an additional WHERE condition to the existing one.
-     * The new condition and the existing one will be joined using the 'AND' operator.
-     * @param string|array $condition the new WHERE condition. Please refer to [[where()]]
-     * on how to specify this parameter.
-     * @param array $params the parameters (name => value) to be bound to the query.
-     * @return static the query object itself
-     * @see where()
-     * @see orWhere()
+     * @inheritdoc
      */
-    public function andWhere($condition, $params = [])
+    public function innerJoin($table, $on = '', $params = [])
     {
-        if ($this->where === null) {
-            $this->where = $condition;
-        } else {
-            $this->where = ['and', $this->where, $condition];
-        }
-        $this->addParams($params);
-        return $this;
+        throw new NotSupportedException('"' . __METHOD__ . '" is not supported.');
     }
 
     /**
-     * Adds an additional WHERE condition to the existing one.
-     * The new condition and the existing one will be joined using the 'OR' operator.
-     * @param string|array $condition the new WHERE condition. Please refer to [[where()]]
-     * on how to specify this parameter.
-     * @param array $params the parameters (name => value) to be bound to the query.
-     * @return static the query object itself
-     * @see where()
-     * @see andWhere()
+     * @inheritdoc
      */
-    public function orWhere($condition, $params = [])
+    public function leftJoin($table, $on = '', $params = [])
     {
-        if ($this->where === null) {
-            $this->where = $condition;
-        } else {
-            $this->where = ['or', $this->where, $condition];
-        }
-        $this->addParams($params);
-        return $this;
+        throw new NotSupportedException('"' . __METHOD__ . '" is not supported.');
     }
 
     /**
-     * Sets the GROUP BY part of the query.
-     * @param string|array $columns the columns to be grouped by.
-     * Columns can be specified in either a string (e.g. "id, name") or an array (e.g. ['id', 'name']).
-     * The method will automatically quote the column names unless a column contains some parenthesis
-     * (which means the column contains a DB expression).
-     * @return static the query object itself
-     * @see addGroupBy()
+     * @inheritdoc
      */
-    public function groupBy($columns)
+    public function rightJoin($table, $on = '', $params = [])
     {
-        if (!is_array($columns)) {
-            $columns = preg_split('/\s*,\s*/', trim($columns), -1, PREG_SPLIT_NO_EMPTY);
-        }
-        $this->groupBy = $columns;
-
-        return $this;
-    }
-
-    /**
-     * Adds additional group-by columns to the existing ones.
-     * @param string|array $columns additional columns to be grouped by.
-     * Columns can be specified in either a string (e.g. "id, name") or an array (e.g. ['id', 'name']).
-     * The method will automatically quote the column names unless a column contains some parenthesis
-     * (which means the column contains a DB expression).
-     * @return static the query object itself
-     * @see groupBy()
-     */
-    public function addGroupBy($columns)
-    {
-        if (!is_array($columns)) {
-            $columns = preg_split('/\s*,\s*/', trim($columns), -1, PREG_SPLIT_NO_EMPTY);
-        }
-        if ($this->groupBy === null) {
-            $this->groupBy = $columns;
-        } else {
-            $this->groupBy = array_merge($this->groupBy, $columns);
-        }
-
-        return $this;
-    }
-
-    /**
-     * Sets the parameters to be bound to the query.
-     * @param array $params list of query parameter values indexed by parameter placeholders.
-     * For example, `[':name' => 'Dan', ':age' => 31]`.
-     * @return static the query object itself
-     * @see addParams()
-     */
-    public function params($params)
-    {
-        $this->params = $params;
-
-        return $this;
-    }
-
-    /**
-     * Adds additional parameters to be bound to the query.
-     * @param array $params list of query parameter values indexed by parameter placeholders.
-     * For example, `[':name' => 'Dan', ':age' => 31]`.
-     * @return static the query object itself
-     * @see params()
-     */
-    public function addParams($params)
-    {
-        if (!empty($params)) {
-            if (empty($this->params)) {
-                $this->params = $params;
-            } else {
-                foreach ($params as $name => $value) {
-                    if (is_integer($name)) {
-                        $this->params[] = $value;
-                    } else {
-                        $this->params[$name] = $value;
-                    }
-                }
-            }
-        }
-
-        return $this;
+        throw new NotSupportedException('"' . __METHOD__ . '" is not supported.');
     }
 
     /**
@@ -758,5 +365,37 @@ class Query extends Component implements QueryInterface
         return $connection->createCommand()
             ->callSnippets($from, $source, $match, $this->snippetOptions)
             ->queryColumn();
+    }
+
+    /**
+     * Creates a new Query object and copies its property values from an existing one.
+     * The properties being copies are the ones to be used by query builders.
+     * @param Query $from the source query object
+     * @return Query the new Query object
+     */
+    public static function create($from)
+    {
+        return new self([
+            'where' => $from->where,
+            'limit' => $from->limit,
+            'offset' => $from->offset,
+            'orderBy' => $from->orderBy,
+            'indexBy' => $from->indexBy,
+            'select' => $from->select,
+            'selectOption' => $from->selectOption,
+            'distinct' => $from->distinct,
+            'from' => $from->from,
+            'groupBy' => $from->groupBy,
+            'join' => $from->join,
+            'having' => $from->having,
+            'union' => $from->union,
+            'params' => $from->params,
+            // Sphinx specifics :
+            'options' => $from->options,
+            'within' => $from->within,
+            'match' => $from->match,
+            'snippetCallback' => $from->snippetCallback,
+            'snippetOptions' => $from->snippetOptions,
+        ]);
     }
 }