+++ framework/web/DbSession.php
@@ -65,6 +65,10 @@ class DbSession extends Session
      *
      * When using DbSession in a production server, we recommend you create a DB index for the 'expire'
      * column in the session table to improve the performance.
+     *
+     * Note that according to the php.ini setting of `session.hash_function`, you may need to adjust
+     * the length of the `id` column. For example, if `session.hash_function=sha256`, you should use
+     * length 64 instead of 40.
      */
     public $sessionTable = '{{%session}}';